+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [x.x.x] - Unreleased
+## [x.x.x] - unreleased
+### Added
+- Support for PUT /sheets/{sheetId}/dataclassification endpoint (Set Data Classification)
+- Support for DELETE /sheets/{sheetId}/dataclassification endpoint (Remove Data Classification)
+- Added `dataClassification` field to Sheet model
 
 ## [4.4.0] - 2026-08-12